Move preflight check into individual scripts
This ensures we don't try to filter NODE_PATH twice, accidentally removing the now-absolute path.
